fix a cast in NodeAnimation

+++ UMI3D-SDK/Assets/UMI3D SDK/ClientDevelopmentKit/Core/Runtime/Animation/UMI3DNodeAnimation.cs
@@ -110,7 +110,7 @@ namespace umi3d.cdk
             switch (property.property)
             {
                 case UMI3DPropertyKeys.AnimationDuration:
-                    ADto.duration = (float)property.value;
+                    ADto.duration = (float)(Double)property.value;
                     break;
                 case UMI3DPropertyKeys.AnimationChain:
                     return UpdateChain(dto, property);
